% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcG.R
\name{dtcG}
\alias{dtcG}
\alias{ptcG}
\alias{qtcG}
\alias{rtcG}
\title{Meta-Gaussian models distributions}
\usage{
dtcG(y, theta, name, ym = 0, step = 0)

ptcG(y, theta, name, ym = 0, step = 0)

qtcG(p, theta, name, ym = 0, step = 0)

rtcG(n, theta, name, ym = 0, step = 0)
}
\arguments{
\item{y}{vector of quantiles (rainfall)}

\item{theta}{model parameters}

\item{name}{name of the anamorphosis, one of \code{'power'},\code{'power-exp'}, \code{'quadratic-power'}, \code{'gp'}}

\item{ym}{minimal value that can be observed}

\item{step}{discretization step}

\item{p}{vector of probabilities}

\item{n}{number of observations to generate}
}
\value{
\code{dtcG} gives the density, \code{ptcG} gives the distribution function, \code{qtcG} gives the quantile function and \code{rtcG} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation of a meta-Gaussian distribution with parameters theta.
}
\details{
Four meta-Gaussian models are available (name argument): \code{'power'},\code{'power-exp'}, \code{'quadratic-power'}, \code{'gp'}
}
\section{Functions}{
\itemize{
\item \code{ptcG}: probability function

\item \code{qtcG}: quantile function

\item \code{rtcG}: random generation function
}}

\examples{
theta=c(-1,0.2,0.9,0.1)
s=seq(0,2,0.01)

sim = rtcG(1e4, theta, "gp")
d = dtcG(s, theta, "gp")
hist(sim, probability=TRUE)
lines(s, d, col=2)
}
