% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_extGP.R
\name{extGP.barplot}
\alias{extGP.barplot}
\title{Barplot for the extended GP model}
\usage{
extGP.barplot(
  y,
  theta,
  ym = 0.2,
  step = 0.2,
  cols = c(1, rgb(0.6, 0.3, 0, 0.5)),
  zoom = NULL,
  legend = TRUE,
  main = "Density of positive low intensities"
)
}
\arguments{
\item{y}{vector of data}

\item{theta}{vector of model parameters}

\item{ym}{minimal value that can be observed}

\item{step}{discretization step}

\item{cols}{vector of length 2, colors for the barplot}

\item{zoom}{vector of length 2, range to zoom on the x axis}

\item{legend}{boolean, should there be a legend?}

\item{main}{title of the plot}
}
\description{
Barplot for the extended GP model
}
