% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_extGP.R
\name{extGP.fit}
\alias{extGP.fit}
\title{Fitting the extended GP model}
\usage{
extGP.fit(
  y,
  init,
  ym = 0,
  step = 0,
  bootstrap = TRUE,
  R = 500,
  plots = TRUE,
  filename = NULL,
  ncpus = parallel::detectCores(),
  ym.param = FALSE
)
}
\arguments{
\item{y}{vector of positive data}

\item{init}{initialisation parameters, a vector of length 3}

\item{ym}{minimal value that can be observed}

\item{step}{discretization step}

\item{bootstrap}{boolean, should boostrap replicates be fitted?}

\item{R}{number of bootstrap replicates}

\item{plots}{boolean, should plots be drawn?}

\item{filename}{if you want to save the plots, a character for your file's name}

\item{ncpus}{number of cores}

\item{ym.param}{boolean, should \code{ym} be a parameter?}
}
\value{
A list of four elements containing the results of the fitted model.
\code{par} gives the fitted parameters.
\code{AIC} gives the Akaike criterion.
\code{fit.boot} gives the parameters fitted with the bootstrap replicates.
Finally \code{for.plots} is a list that is used by the function \code{res.plot}.
}
\description{
Fitting the extended GP model
}
\details{
The extended GP model \insertCite{naveau2016modeling}{tcG} is define for positive rainfall as

\eqn{Y+ = ym + \sigma H_\xi^{-1}(U^{1/\alpha})},

with \eqn{H_\xi} the cdf of a GPD.
}
\examples{
\dontrun{
res2=extGP.fit(rain[rain>0], c(2,.5,.5), ym=.3, step=.1, R=50)
}
}
\references{
{
\insertAllCited{}
}
}
