% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot_extGP.R
\name{extGP.qqplot}
\alias{extGP.qqplot}
\title{QQplot for the extended GP model}
\usage{
extGP.qqplot(
  y,
  qfit,
  q.L,
  q.U,
  cols = c(rgb(0.6, 0.3, 0), rgb(0.6, 0.3, 0, 0.5), rgb(0.6, 0.3, 0, 0.1)),
  zoomx = range(y, na.rm = T),
  zoomy = range(y, na.rm = T),
  legend = TRUE,
  main = "QQ plot of positive rainfall",
  add = FALSE
)
}
\arguments{
\item{y}{vector of data}

\item{qfit}{list obtained with \code{fit.extGP} (fitted quantiles)}

\item{q.L}{list obtained with \code{fit.extGP} (lower intervall of the bootstrap quantiles)}

\item{q.U}{list obtained with \code{fit.extGP} (upper intervall of the bootstrap quantiles)}

\item{cols}{vector of colors for the barplot}

\item{zoomx}{vector of length 2, range to zoom on the x axis}

\item{zoomy}{vector of length 2, range to zoom on the y axis}

\item{legend}{boolean, should there be a legend?}

\item{main}{title of the plot}

\item{add}{boolean, should the plot be added to the current plot?}
}
\description{
Called by \code{res.plot}, we don't advise you to use it.
}
