% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_choice.R
\name{ll.choice}
\alias{ll.choice}
\title{Log-likelihood function of a meta-Gaussian model}
\usage{
ll.choice(discrete, name = "power-exp")
}
\arguments{
\item{discrete}{boolean. \code{TRUE} for the discrete likelihhod, \code{FALSE} for the continuous one}

\item{name}{name of the anamorphosis, one of \code{'power'},\code{'power-exp'}, \code{'quadratic-power'}, \code{'gp'}}
}
\value{
The log-likelihood function of meta-Gaussian model with anamorphosis \code{name}.
}
\description{
Log-likelihood function of a meta-Gaussian model
}
\examples{
ll.choice(TRUE, "power")
}
