% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{momtcG}
\alias{momtcG}
\alias{momtcG.pos}
\title{Moments of GP meta-Gaussian model}
\usage{
momtcG(p, mu, sigma, alpha, xi, ym = 0)

momtcG.pos(p, mu, sigma, alpha, xi)
}
\arguments{
\item{p}{order of the moment}

\item{mu}{mean of the Gaussian}

\item{sigma}{scale parameter}

\item{alpha}{lower tail parameter}

\item{xi}{upper tail parameter}

\item{ym}{minimal value that can be observed}
}
\value{
Moment of order p of the GP meta-Gaussian distribution
}
\description{
Computes the moments of the GP meta-Gaussian model, all distribution (momtcG) and positive part (momtcG.pos)
}
\section{Functions}{
\itemize{
\item \code{momtcG.pos}: Moment of the positive part of the GP meta-Gaussian distribution
}}

