% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{Anamorphosis function}
\usage{
psi(theta, x, ym = 0, name = "power-exp")
}
\arguments{
\item{theta}{model parameters}

\item{x}{value to apply psi (Gaussian)}

\item{ym}{minimal value that can be observed}

\item{name}{name of the anamorphosis, one of \code{'power'},\code{'power-exp'}, \code{'quadratic-power'}, \code{'gp'}}
}
\value{
Precipitation amount from transforming x
}
\description{
A function that computes the anamorphosis of four meta-Gaussian models (see Details)
}
\examples{
psi(c(-1,.2,.6,.1), 0.87, name='gp')
}
