% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi1.R
\name{psi1}
\alias{psi1}
\title{Inverse of the anamorphosis}
\usage{
psi1(theta, y, ym = 0, name = "power-exp")
}
\arguments{
\item{theta}{model paramters}

\item{y}{value to apply psi^{-1} (rainfall)}

\item{ym}{minimal value that can be observed}

\item{name}{name of the anamorphosis, one of \code{'power'},\code{'power-exp'}, \code{'quadratic-power'}, \code{'gp'}}
}
\value{
Returns the inverse of psi
}
\description{
A function that computes the inverse of the anamorphosis of four meta-Gaussian models (see Details)
}
\examples{
psi1(c(-1,.2,.6), 0.1, name='power')
}
