% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_res.R, R/res.plot.R
\name{res.plot}
\alias{res.plot}
\title{Plotting the results}
\usage{
res.plot(
  res.tcG = NULL,
  res.extGP = NULL,
  y,
  ym = 0,
  step = 0,
  zoom = range(y, na.rm = TRUE),
  select = NULL,
  choice = c("dens", "qqplot"),
  legend = TRUE,
  ...
)

res.plot(
  res.tcG = NULL,
  res.extGP = NULL,
  y,
  ym = 0,
  step = 0,
  zoom = range(y, na.rm = TRUE),
  select = NULL,
  choice = c("dens", "qqplot"),
  legend = TRUE,
  ...
)
}
\arguments{
\item{res.tcG, res.extGP}{a list resulting respectively from \code{fit.tcG} and \code{fit.extGP}, at least one of them must be non-null}

\item{y}{vector of data (if both \code{res.tcG} and \code{res.extGP} are given, \code{y} must include the zeros)}

\item{ym}{minimal value that can be observed}

\item{step}{discretization step}

\item{zoom}{vector of length 2, range to zoom on}

\item{select}{for \code{res.tcG}, a the model(s) that should be plotted. A subset of \code{c('gp', 'power', 'power-exp', 'quadratic-power')}}

\item{choice}{A subset of \code{c('qqplot', 'dens')}. Which plots should be drawn?}

\item{legend}{boolean, should there be a legend?}

\item{...}{graphical parameters such as \code{main}, \code{cex}, etc.}
}
\description{
This function offers choices for plotting results of \code{fit.tcG} and \code{fit.extGP}.
Two types of plots are available: a density plot (histogram or barplot) and a QQ plot.
In particular the function allows plotting the results of both fit functions together on the QQ plot.

This function offers choices for plotting results of \code{fit.tcG} and \code{fit.extGP}.
Two types of plots are available: a density plot (histogram or barplot) and a QQ plot.
In particular the function allows plotting the results of both fit functions together on the QQ plot.
}
\examples{
\dontrun{

# Comparing meta-Gaussian models

data(rain, package = "ismev")

res=tcG.fit(rain, name=c("power", "gp"),
            init=list("power"=c(0,2,2), "gp"=c(0,2,0.5,0.5)),
            ym=0.3, step=0.1, R=50, plots=FALSE)
res.plot(res.tcG=res, y=rain, ym=.3, step=.1, zoom=c(0,30), choice="qqplot")

# Comparing the GP meta-Gaussian and the extended GP

res2=extGP.fit(rain[rain>0], c(2,.5,.5), ym=.3, step=.1, R=50, plots=FALSE)
res.plot(res.tcG=res, res.extGP=res2, y=rain, ym=.3, step=.1, select="gp", choice="qqplot")

}
\dontrun{

# Comparing meta-Gaussian models

data(rain, package = "ismev")

res=tcG.fit(rain, name=c("power", "gp"),
            init=list("power"=c(0,2,2), "gp"=c(0,2,0.5,0.5)),
            ym=0.3, step=0.1, R=50, plots=FALSE)
res.plot(res.tcG=res, y=rain, ym=.3, step=.1, zoom=c(0,30), choice="qqplot")

# Comparing the GP meta-Gaussian and the extended GP

res2=extGP.fit(rain[rain>0], c(2,.5,.5), ym=.3, step=.1, R=50, plots=FALSE)
res.plot(res.tcG=res, res.extGP=res2, y=rain, ym=.3, step=.1, select="gp", choice="qqplot")

}
}
