% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_tcG.R
\name{tcG.barplot}
\alias{tcG.barplot}
\title{Barplot for meta-Gaussian fitted model}
\usage{
tcG.barplot(
  y,
  theta,
  name = "gp",
  ym = 0.2,
  step = 0.2,
  cols = c(1, rgb(0, 0, 1, 0.5)),
  zoom = NULL,
  legend = TRUE,
  main = "Density of low intensities"
)
}
\arguments{
\item{y}{vector of data}

\item{theta}{vector of model parameters}

\item{name}{name of the anamorphosis, one of \code{'power'},\code{'power-exp'}, \code{'quadratic-power'}, \code{'gp'}}

\item{ym}{minimal value that can be observed}

\item{step}{discretization step}

\item{cols}{vector of length 2, colors for the barplot}

\item{zoom}{vector of length 2, range to zoom on the x axis}

\item{legend}{boolean, should there be a legend?}

\item{main}{title of the plot}
}
\description{
Called by \code{res.plot}, we don't advise you to use it.
}
