% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot_tcG.R
\name{tcG.qqplot}
\alias{tcG.qqplot}
\title{QQplot for meta-Gaussian models}
\usage{
tcG.qqplot(
  y,
  names,
  qfit,
  q.L,
  q.U,
  cols = cbind(c(rgb(0, 0, 1), rgb(1, 0, 0), rgb(0, 1, 0), rgb(1, 0, 1)), c(rgb(0, 0,
    1, 0.5), rgb(1, 0, 0, 0.5), rgb(0, 1, 0, 0.5), rgb(1, 0, 1, 0.5)), c(rgb(0, 0, 1,
    0.1), rgb(1, 0, 0, 0.1), rgb(0, 1, 0, 0.1), rgb(1, 0, 1, 0.1))),
  zoomy = range(y, na.rm = T),
  legend = TRUE,
  zoomx = range(y, na.rm = T),
  main = "QQ plot"
)
}
\arguments{
\item{y}{vector of data}

\item{names}{name(s) of the anamorphosis that were fitted, a subset of \code{c('gp', 'power', 'power-exp', 'quadratic-power')}}

\item{qfit}{list obtained with \code{fit.tcG} (fitted quantiles)}

\item{q.L}{list obtained with \code{fit.tcG} (lower intervall of the bootstrap quantiles)}

\item{q.U}{list obtained with \code{fit.tcG} (upper intervall of the bootstrap quantiles)}

\item{cols}{matrix of colors for the barplot}

\item{zoomy}{vector of length 2, range to zoom on the y axis}

\item{legend}{boolean, should there be a legend?}

\item{zoomx}{vector of length 2, range to zoom on the x axis}

\item{main}{title of the plot}
}
\description{
Called by \code{res.plot}, we don't advise you to use it.
}
